[front] chore: move consumption scope contracts to types
