+++ front/components/workspace/analytics/consumption/ConsumptionAttributionBreakdown.tsx
@@ -1,8 +1,8 @@
 import type { ConsumptionTopRow } from "@app/hooks/useConsumptionTop";
 import { useConsumptionTop } from "@app/hooks/useConsumptionTop";
 import type { ConsumptionPeriodSelection } from "@app/lib/analytics/consumption_period";
-import type { ConsumptionScopeFilter } from "@app/lib/api/analytics/consumption/scope";
-import { CONSUMPTION_DIMENSION_FILTER_KEYS } from "@app/lib/api/analytics/consumption/scope";
+import type { ConsumptionScopeFilter } from "@app/types/api/analytics/consumption";
+import { CONSUMPTION_DIMENSION_FILTER_KEYS } from "@app/types/api/analytics/consumption";
 import { Button, cn, LoadingBlock, ProgressBar } from "@dust-tt/sparkle";
 import type { ComponentType } from "react";
 import type { ConsumptionDimension } from "./consumptionDimensions";
